initiate new db on no-file open
When flatfile.db is called on a file that does not exist, we should
create a new object and return it rather than returning an error.

This allows the user to create a new database file with flatfile.db
without having to explicitly create it and populate it beforehand.

This is a semver-major change; before this, flatfile.db would give an
ENOENT error when called on a file that does not exist.

Semver: major
Fixes: https://github.com/brendanashworth/flatfile/issues/2